Refactor nemo predictor code into base class
